Add event to create diagram
