+++ js/views/root.js
@@ -8,20 +8,10 @@ Backbone.$ = $;
 module.exports = Backbone.View.extend({
   template: _.template($('#rootTemplate').html()),
 
-  events: {
-    "click button": "createNew",
-  },
-
   initialize: function(){
     _.bindAll(this, 'render');
 
     this.render();
-  },
-
-  createNew: function(event) {
-    event.preventDefault();
-    var modelId = $('input#modelId').val();
-    Backbone.history.navigate("/" + modelId, {trigger: true});
   },
 
   renderDiagrams: function(content) {
